requirements.txt
pymupdf
qdrant-client
sentence-transformers
fastembed
google-generativeai
streamlit
python-dotenv